Add files via upload
v1.03

Bugfixes:

Fixed a bug where Hitokotonusi was named Hitoko-Nushi. (For the record, the second spelling is the correct one, but it turns out that the English version of the game uses the first spelling.) Also solved an issue where some Persona calculations were incorrect because of a small bug in the level calculator where I forgot to account for how the game rounds a result with a decimal. (8/9/22)